requirements.txt
asgiref==3.7.2
Django==4.2.4
django-tenant-schemas==1.11.0
ordered-set==4.1.0
postgres==4.0
psycopg2-binary==2.9.7
psycopg2-pool==1.1
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
